Moved enums to base class to reduce circular dependencies
