+++ src/raspberrypi/controllers/sasidev_ctrl.h
@@ -30,6 +30,74 @@
 //===========================================================================
 class SASIDEV
 {
+protected:
+	enum scsi_message_code : BYTE {
+		eMsgCodeAbort                              = 0x06,                           
+		eMsgCodeAbortTag                           = 0x0D,
+		eMsgCodeBusDeviceReset                   = 0x0C,
+		eMsgCodeClearQueue             = 0x0E,
+		eMsgCodeCommandComplete                   = 0x00,
+		eMsgCodeDisconnect                         = 0x04,
+		eMsgCodeIdentify                           = 0x80,
+		eMsgCodeIgnoreWideResidue    = 0x23, // (Two Bytes)
+		eMsgCodeInitiateRecovery                  = 0x0F,
+		eMsgCodeInitiatorDetectedError           = 0x05,
+		eMsgCodeLinkedCommandComplete            = 0x0A,
+		eMsgCodeLinkedCommandCompleteWithFlag = 0x0B,
+		eMsgCodeMessageParityError               = 0x09,
+		eMsgCodeMessageReject                     = 0x07,
+		eMsgCodeNoOperation                       = 0x08,
+		eMsgCodeHeadOfQueueTag                = 0x21,
+		eMsgCodeOrderedQueueTag                = 0x22,
+		eMsgCodeSimpleQueueTag                 = 0x20,
+		eMsgCodeReleaseRecovery                   = 0x10,
+		eMsgCodeRestorePointers                   = 0x03,
+		eMsgCodeSaveDataPointer                  = 0x02,
+		eMsgCodeTerminateIOProcess              = 0x11,
+	};
+
+	enum scsi_command : BYTE {
+			eCmdTestUnitReady = 0x00,
+			eCmdRezero =  0x01,
+			eCmdRequestSense = 0x03,
+			eCmdFormat = 0x04,
+			eCmdReassign = 0x07,
+			eCmdRead6 = 0x08,
+			eCmdRetrieveStats = 0x09,    // DaynaPort specific command
+			eCmdWrite6 = 0x0A,
+			eCmdSeek6 = 0x0B,
+			eCmdSetIfaceMode = 0x0C,     // DaynaPort specific command
+			eCmdSetMcastAddr  = 0x0D,    // DaynaPort specific command
+			eCmdEnableInterface = 0x0E,  // DaynaPort specific command
+			eCmdInquiry = 0x12,
+			eCmdModeSelect = 0x15,
+			eCmdReserve6 = 0x16,
+			eCmdRelease6 = 0x17,
+			eCmdModeSense = 0x1A,
+			eCmdStartStop = 0x1B,
+			eCmdRcvDiag = 0x1C,
+			eCmdSendDiag = 0x1D,
+			eCmdRemoval = 0x1E,
+			eCmdReadCapacity = 0x25,
+			eCmdRead10 = 0x28,
+			eCmdWrite10 = 0x2A,
+			eCmdSeek10 = 0x2B,
+			eCmdWriteAndVerify10 = 0x2E,
+			eCmdVerify = 0x2F,
+			eCmdSynchronizeCache = 0x35,
+			eCmdReadDefectData10 = 0x37,
+			eCmdReadToc = 0x43,
+			eCmdPlayAudio10 = 0x45,
+			eCmdPlayAudioMSF = 0x47,
+			eCmdPlayAudioTrack = 0x48,
+			eCmdModeSelect10 = 0x55,
+			eCmdReserve10 = 0x56,
+			eCmdRelease10 = 0x57,
+			eCmdModeSense10 = 0x5A,
+			eCmdInvalid = 0xC2,		// (SASI only/Suppress warning when using SxSI)
+			eCmdSasiCmdAssign = 0x0e, // This isn't used by SCSI, and can probably be removed.
+	};
+
 public:
 	enum {
 		UnitMax = 8					// Maximum number of logical units
